requirements.txt
# ResearchHub AI - MVP Requirements

# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23

# WebSocket
python-socketio==5.10.0
python-engineio==4.8.0

# Security
bcrypt==4.1.2
Werkzeug==3.0.1

# AI/LLM Integration
openai==1.6.1
langchain==0.1.0
langchain-community==0.0.13

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0

# Development
python-dateutil==2.8.2

# PDF & Document Generation
reportlab==4.0.7
PyPDF2==3.0.1
markdown2==2.4.12
weasyprint==60.1
Jinja2==3.1.3
